+++ src/main/java/de/samply/query/OutputFormat.java
@@ -1,0 +1,10 @@
+package de.samply.query;
+
+public enum OutputFormat {
+    CSV,
+    EXCEL,
+    OPAL,
+    JSON,
+    XML;
+
+}