+++ src/main/java/de/samply/query/QueryProcess.java
@@ -1,0 +1,7 @@
+package de.samply.query;
+
+public enum QueryProcess {
+    EXPORT,
+    DATASHIELD,
+    DATA_SCIENCE_ORCHESTRATOR
+}