+++ src/main/resources/db/migration/V001__initialize_schema_and_tables.sql
@@ -45,6 +45,27 @@ CREATE TABLE samply.project_manager_admin_user
     email TEXT
 );
 
+CREATE TABLE samply.project_document
+(
+    id         SERIAL NOT NULL PRIMARY KEY,
+    file_path  TEXT,
+    url        TEXT,
+    project_id BIGINT NOT NULL
+);
+
+CREATE TABLE samply.project_query
+(
+    id            SERIAL PRIMARY KEY,
+    project_id    BIGINT NOT NULL,
+    query         TEXT   NOT NULL,
+    query_format  TEXT   NOT NULL,
+    output_format TEXT   NOT NULL,
+    template_id   TEXT   NOT NULL,
+    label         TEXT,
+    description   TEXT,
+    query_process TEXT   NOT NULL
+);
+
 ALTER TABLE samply.project_bridgehead
     ADD CONSTRAINT fk_project_id
         FOREIGN KEY (project_id)
@@ -55,5 +76,17 @@ ALTER TABLE samply.project_bridgehead_user
         FOREIGN KEY (project_bridgehead_id)
             REFERENCES samply.project_bridgehead (id);
 
+ALTER TABLE samply.project_document
+    ADD CONSTRAINT fk_project_document_project
+        FOREIGN KEY (project_id)
+            REFERENCES samply.project (id);
+
+ALTER TABLE samply.project_query
+    ADD CONSTRAINT fk_project_query_project
+        FOREIGN KEY (project_id)
+            REFERENCES samply.project (id);
+
 CREATE INDEX idx_project_bridgehead_project_id ON samply.project_bridgehead (project_id);
 CREATE INDEX idx_project_bridgehead_user_project_bridgehead_id ON samply.project_bridgehead_user (project_bridgehead_id);
+CREATE INDEX idx_project_document_project_id ON samply.project_document (project_id);
+CREATE INDEX idx_project_query_project_id ON samply.project_query (project_id);